# platform/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OfertaItemResumo": {
        "properties": {
          "altura": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altura"
          },
          "criado_em": {
            "format": "date-time",
            "title": "Criado Em",
            "type": "string"
          },
          "de_nome": {
            "title": "De Nome",
            "type": "string"
          },
          "de_personagem_id": {
            "title": "De Personagem Id",
            "type": "string"
          },
          "estado": {
            "enum": [
              "pendente",
              "aceita",
              "recusada",
              "cancelada"
            ],
            "title": "Estado",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "item_nome": {
            "title": "Item Nome",
            "type": "string"
          },
          "largura": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Largura"
          },
          "para_nome": {
            "title": "Para Nome",
            "type": "string"
          },
          "para_personagem_id": {
            "title": "Para Personagem Id",
            "type": "string"
          },
          "subtipo": {
            "anyOf": [
              {
                "enum": [
                  "peitoral",
                  "capacete",
                  "luvas",
                  "botas",
                  "uma_mao",
                  "duas_maos",
                  "escudo",
                  "mochila",
                  "aljava",
                  "moedas",
                  "outro"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtipo"
          }
        },
        "required": [
          "id",
          "estado",
          "item_id",
          "item_nome",
          "de_personagem_id",
          "de_nome",
          "para_personagem_id",
          "para_nome",
          "criado_em"
        ],
        "title": "OfertaItemResumo",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API autoritativa da plataforma colaborativa Cursed.",
    "title": "Cursed — Plataforma Colaborativa",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/mesas/{mesa_id}/ofertas-item": {
      "get": {
        "description": "Ofertas pendentes em que o usuário pode agir: dadas ou recebidas por personagens que ele edita.",
        "operationId": "listar_ofertas_mesas__mesa_id__ofertas_item_get",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OfertaItemResumo"
                  },
                  "title": "Response Listar Ofertas Mesas  Mesa Id  Ofertas Item Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar Ofertas",
        "tags": [
          "Sala"
        ]
      }
    },
    "/mesas/{mesa_id}/ofertas-item/{oferta_id}/aceitar": {
      "post": {
        "operationId": "aceitar_oferta_mesas__mesa_id__ofertas_item__oferta_id__aceitar_post",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "oferta_id",
            "required": true,
            "schema": {
              "title": "Oferta Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AceitarOfertaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemInventarioResumo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Aceitar Oferta",
        "tags": [
          "Sala"
        ]
      }
    },
    "/mesas/{mesa_id}/ofertas-item/{oferta_id}/cancelar": {
      "post": {
        "operationId": "cancelar_oferta_mesas__mesa_id__ofertas_item__oferta_id__cancelar_post",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "oferta_id",
            "required": true,
            "schema": {
              "title": "Oferta Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaItemResumo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancelar Oferta",
        "tags": [
          "Sala"
        ]
      }
    },
    "/mesas/{mesa_id}/ofertas-item/{oferta_id}/recusar": {
      "post": {
        "operationId": "recusar_oferta_mesas__mesa_id__ofertas_item__oferta_id__recusar_post",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "oferta_id",
            "required": true,
            "schema": {
              "title": "Oferta Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaItemResumo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recusar Oferta",
        "tags": [
          "Sala"
        ]
      }
    },
...
    "/mesas/{mesa_id}/personagens/{personagem_id}/inventario/{item_id}/largar": {
      "post": {
        "operationId": "largar_item_mesas__mesa_id__personagens__personagem_id__inventario__item_id__largar_post",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LargarItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipienteResumo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Largar Item",
        "tags": [
          "Sala"
        ]
      }
    },
    "/mesas/{mesa_id}/personagens/{personagem_id}/inventario/{item_id}/ofertas": {
      "post": {
        "operationId": "ofertar_item_mesas__mesa_id__personagens__personagem_id__inventario__item_id__ofertas_post",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfertarItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaItemResumo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ofertar Item",
        "tags": [
          "Sala"
        ]
      }
    },
...
    "/mesas/{mesa_id}/sala/recipientes": {
      "get": {
        "description": "Chão e baús da cena ativa, visíveis a todos os participantes.",
        "operationId": "listar_recipientes_mesas__mesa_id__sala_recipientes_get",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecipienteResumo"
                  },
                  "title": "Response Listar Recipientes Mesas  Mesa Id  Sala Recipientes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar Recipientes",
        "tags": [
          "Sala"
        ]
      },
      "post": {
        "operationId": "criar_bau_mesas__mesa_id__sala_recipientes_post",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarBauRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipienteResumo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Criar Bau",
        "tags": [
          "Sala"
        ]
      }
    },
...
  }
}